server: assign a random ID to File's without one during Create
Issue: https://github.com/moov-io/ach/issues/401
